game_info/requirements.txt
django
python-valve
djangorestframework
